Fix handling of Erubis templates with xss escaping
either with rails_xss or Rails 3.
This was broken when Brakeman's Erubis output was
changed to match what rails_xss does. Unfortunately, that
broke the ErubisTemplateProcessor such that NO output
was detected. This should fix that.

Note that this code detects auto-escaping by the output variable.
@output_buffer is used in Brakeman's Erubis classes. _buf will
only show up if someone is using Erubis with auto-escaping turned
off.