Fix duplicate warnings with `raw` calls
There was a bit of a logic error here, because a
warning might be detected as a duplicate in
`check_for_immediate_xss`, causing it to return 'false'
and then the check would go ahead and process the `raw` call.
This would result in 'weak' confidence warnings that were
duplicates of 'high' confidence warnings.