Changes "exit_on_warn" option.
Previously, "exit on warn" would return a non-zero exit code when warnings were fixed, which was not desirable when using brakeman on a CI system that depends on the exit status to determine if new vulnerabilities have been detected.

Now brakeman will only "fail" a build when a warning has been introduced, not when it has been resolved.