Expand slim version support to latest.
At the time of this commit that's slim v3.0.1.

That is, except for Ruby 1.8.7 installations. Slim dropped support[^0] for Ruby
1.8.7 in their 3.0.0 release. This commit allows Brakeman to move forward with
slim support on all supported later Ruby versions.

To do this accurately we must install an appropriate version of Slim by
inspecting `RUBY_VERSION` at install time. We do that by using the native
extensions hooks available as part of gem installs. Ours is modeled after an
example on GitHub[^1].

[^0]: https://github.com/slim-template/slim/blob/master/CHANGES#L12
[^1]: https://github.com/mmzyk/gem_dependency_example/tree/master/gem_dep_example