Use Warning#link when determining local path to warning doc content
To support documentation content (readups) in the Code Climate Platform,
markdown content was synced from the presidentbeef/brakeman-site
repository and included in the [initial PR][cc-support] to add Code
Climate support.

Currently, the Code Climate reporter generates a local path using the
warning's type, which doesn't always match the local directory holding
the documentation content. What we get instead is the fallback content
of "Read more: #{link}", even though documentation is available.  Checks
often provide a full `link_path` and `Warning#link` already contains
logic for generating the correct url for documentation, so we can defer
to it when computing a local path to check.

[cc-support]: presidentbeef/brakeman#758