Fix CodeClimate construction of --only-files
Our construction of this string was naive and didn't account for path
names that might contain characters like spaces. This would result in
`bin/brakeman` getting arguments that would be parsed as pointing at a
completely invalid app root.

By shell-escaping the entire string of paths, we can avoid this.