Key instance and class methods differently
If a collection had a class method and an instance method that shared
the same name, one would overwrite the other since they are keyed by
name only.

This change keys the class methods prefixed with this class. Since the
tracker does this in a couple of different places anyway, it can now
just rely on the key to return the method name.

Since fingerprints are based partly on this key, a few tests had to be
updated that checked for this.

Fixes https://github.com/presidentbeef/brakeman/issues/880