Raise errors for files with commas in their name
We received a report that analysis was failing for a repo and tracked it
down to the fact that there was a file in the repo whose filename ended
in a comma. This leads to a fairly confusing error.

This change surfaces an error which explains to the user what is going
on, and what they can do about it.

If there is a comma in the middle of the filename, that doesn't lead to
a runtime error currently, but it does lead to misbehavior, so I think
an error is also appropriate for those files.