Code Climate engine: fix `include_paths` when using custom app path
Brakeman (very reasonably) expects the `only_files` options to be
relative to the `app_path` option. The support that was added to the
Code Climate engine for a custom `app_path` option did not take
`include_paths` into account, so the normal result of using a custom
`app_path` would have ended up being that *nothing* would get analyzed:
because with a custom `app_path` of, say, `foo`, an include path might
include something like `foo/bar` when it needs to be adjusted to just
`bar`.

The fix here is to strip the `app_path` prefix from all matching
`include_paths`, and drop non-matching `include_paths` entirely.