Code Climate engine: handle deeply nested app paths
If an `app_path` is *below* the expanded `include_paths`, the simple
approach of "delete the prefix from the include paths" would incorrectly
result in no include paths at all.

The fix here is, when stripping the include paths, to consider a path
that is one of the `app_paths` direct ancestors to map to "." for the
include paths.