Code Climate: omit leading dot from `only_files`
When an `app_path` is specified in the Code Climate configuration, the
relative paths that
`Brakeman::Codeclimate::EngineConfiguration#active_include_paths`
generates (added in presidentbeef/brakeman#1126) all start with a
dot-slash (ex. `./app/`). This unfortunately causes a bug downstream in
`Brakeman::AppTree#select_only_files` where _no_ paths actually match
the `@only_files` regular expression.

A potential reason we missed this bug in the first place is the
repository we were testing with used an `app_path` of `app`, which
outwardly appears to mostly work.