mv codeclimate dockerfile->Dockerfile.codeclimate
This restores the functionality of the codeclimate Dockerfile, but names
it Dockerfile.codeclimate so that the user can use the default
Dockerfile to run brakeman with access to the full set of command-line
arguments.

[Metadata: I could have rewritten history to `git mv` the original
Dockerfile -> Dockerfile.codeclimate and then added the new default
Dockerfile, but I felt it was easier to just re-add the original file
under this new name. Hopefully the commit history isn't too confusing.]