Change deprecated MAINTAINER to use label
From https://docs.docker.com/engine/reference/builder/#maintainer-deprecated:

The MAINTAINER instruction sets the Author field of the generated images. The LABEL instruction is a much more flexible version of this
and you should use it instead, as it enables setting any metadata you require, and can be viewed easily, for example with docker inspect.

Meta-note: We might want to add the associated e-mail instead of just first+last name