Skip dynamic render path check for renderables
[Rails 6.1 now supports passing `Renderable`s to `render`](https://github.com/rails/rails/blob/b5b078cd99b0107fd8e31ff619578e5321182e0a/actionview/lib/action_view/template/renderable.rb).

A renderable is anything that responds to `render_in`. This means we can
now expect to see `render(SomeRenderableClass.new(some_parameter))`. A
new library for building Rails UIs out of components, ViewComponent, is
growing in popularity and makes use of this new `render_in` API.

Special handling is included for detecting descendants of
`ViewComponent::Base` so that dependencies don't need to be indexed.