Don't exclude requested engine or lib paths if they reside in vendor
Fixes #1679

Currently, we find files that match various patterns such as the engine or
lib paths and exclude them if they reside in a vendor directory.

Instead, we should honor explicitly-allowed paths, even if we're excluding
vendor and they reside in a vendor directory.

Note, there are several globally EXCLUDED_PATHS such as:
/generators, test/, log/ that are still excluded even if they're in opt-in
engine/lib paths. This still seems reasonable and a test is included to
demonstrate this behavior is unchanged.