Fix compatibility with `--enable-frozen-string-literal`
Since Ruby 2.3, Ruby can be started with this options, and it's
expected to become the default in Ruby 4.0 (no release date yet).

This commit fixes the callsites that assume string literal are mutable.

Note however that a large part of the test suite is still failing,
because of two dependencies:

  - ruby2ruby: https://github.com/seattlerb/ruby2ruby/pull/58
  - erubis: (this one is abandoned so can hardly be fixed).