Check for unsafe SQL when two arguments are passed to AR methods
I think this conditional is meant to

1. Always check the first and last arguments.
2. Avoid checking the argument twice when there's only one argument.

The structure of `call` appears to be:

* 0: the symbol `:call`
* 1: the receiver of the method call
* 2: the method being called
* 3: the first argument
* 4: the second argument
* 5: the third argument (and so on)

So `call.length > 5` will be true only if there are at least 3
arguments.

Before this change, the behavior was

```ruby
User.count("#{params[:input]}") # triggers warning
User.count("#{params[:input]}", :foo) # doesn't trigger warning
User.count("#{params[:input]}", :foo, :bar) # triggers warning
```

After this change, the two-argument version will trigger the warning.

Another option would be to use `if call.arglist.length > 1`. This is
more explicit, but it involves an additional method call, and I don't
know what the performance implications of that are.